requirements.txt
crewai==0.86.0
python-dotenv==1.0.1